% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_CDS.R
\name{extract_CDSfrom_xm}
\alias{extract_CDSfrom_xm}
\title{Extract the coding sequences (CDS) coordinates from a XM accession}
\usage{
extract_CDSfrom_xm(listName)
}
\arguments{
\item{listName}{a downloaded flat file from the nuccore NCBI database}
}
\value{
An `IRanges` object with the start and end position of the CDS of the
putative mRNA.
}
\description{
Parses an XM accession (Genbank format) and extract the CDS coordinates.
The CDS coordinates refer to the mRNA molecule.
}
\examples{
xm <- "XM_020388824"
# First, get the character vector containing the fetched record
target <- rentrez::entrez_fetch(db = "nuccore", id = xm, rettype = "gp")
extract_CDSfrom_xm(target)
}
\author{
Jose V. Die
}
