% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mol.wt.R
\name{extract_mol.wt_from_xp}
\alias{extract_mol.wt_from_xp}
\title{Extract the molecular weigth from an XP accession}
\usage{
extract_mol.wt_from_xp(listName, feat = "calculated_mol_wt")
}
\arguments{
\item{listName}{a downloaded flat file from the protein NCBI database}

\item{feat}{a feature to be extracted. Allowed feature is only "calculated_mol_wt"}
}
\description{
Parses an XP acession output (Genbank format) and extract the molecular weight
(in Daltons). First it is needed to fecth the protein id info (using the
'rentrez' packg). Then, the function takes that output as an argument.
}
\details{
First, get the character vector containing the fetched record. Then, this
function parses the fectched record and returns the molecular weight.
}
\examples{
xp <- "XP_020244413"
# First, get the character vector containing the fetched record
protein_gb <- rentrez::entrez_fetch(db = "protein", id = xp, rettype = "gp")
extract_mol.wt_from_xp(protein_gb)
}
\seealso{
\code{\link[rentrez]{entrez_fetch}}
}
\author{
Jose V. Die
}
