% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXM.R
\name{getXM}
\alias{getXM}
\title{Get the XM accession from XP accession}
\usage{
getXM(xp)
}
\arguments{
\item{xp}{a vector containing a single XP id.}
}
\description{
Returns the XM accession from a single XP accession.
}
\examples{
xp <- "XP_020244413"
getXM(xp)
}
\author{
Jose V. Die
}
