% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXMfromLOC.R
\name{getXMfromLOC}
\alias{getXMfromLOC}
\title{Get the XM accession from LOC symbol}
\usage{
getXMfromLOC(locIDs)
}
\arguments{
\item{locIDs}{a characyer vector containing a single / multiple LOC symbols.}
}
\description{
Returns a character vector containing the XM accession from a from a single /
multiple LOC symbols.
}
\examples{
locIDs <- c("LOC101515097", "LOC101515098", "LOC101515099")
getXMfromLOC(locIDs)
}
\author{
Jose V. Die
}
