% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXP.R
\name{getXP}
\alias{getXP}
\title{Get the XP accession from XM accession}
\usage{
getXP(xm)
}
\arguments{
\item{xm}{a vector containing a single XM id.}
}
\description{
Returns the XP accession from a single XM accession.
}
\examples{
xm <- "XM_020388824"
getXP(xm)
}
\author{
Jose V. Die
}
