% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXPfromLOC.R
\name{getXPfromLOC}
\alias{getXPfromLOC}
\title{Get the XP accession from LOC symbol}
\usage{
getXPfromLOC(locIDs)
}
\arguments{
\item{locIDs}{a characyer vector containing a single / multiple LOC symbols.}
}
\description{
Returns a character vector containing the XP accession from a from a single /
multiple LOC symbols.
}
\examples{
locIDs = c("LOC101515097", "LOC101515098", "LOC101515099")
getXPfromLOC(locIDs)

}
\author{
Jose V. Die
}
