% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refseq_fromGene_action.R
\name{refseq_fromGene_action}
\alias{refseq_fromGene_action}
\title{Get the mRNA or protein accession}
\usage{
refseq_fromGene_action(GeneID, sequence, retries)
}
\arguments{
\item{GeneID}{A character string of the GeneID.}

\item{sequence}{A character string of the mRNA or protein accession to fetch data from mRNA or protein databases, respectively.}

\item{retries}{A numeric value to control the number of retry attempts to handle 502 errors.}
}
\value{
A character vector containing the mRNA or protein accession corresponding to the especified \code{GeneID}.
}
\description{
\code{refseq_fromGene_action()} Returns the mRNA or protein accession from a single GeneID.
}
\author{
Jose V. Die
}
