% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refseq_geneSymbol_action.R
\name{refseq_geneSymbol_action}
\alias{refseq_geneSymbol_action}
\title{Get the gene symbol}
\usage{
refseq_geneSymbol_action (id, db, retries)
}
\arguments{
\item{id}{A character string of the transcript or protein id.}

\item{db}{A character string of the "nuccore" or "protein" database.}

\item{retries}{A numeric value to control the number of retry attempts to handle internet errors.}
}
\value{
A character vector containing the gene symbol corresponding to the especified accession as \code{id}.
}
\description{
\code{refseq_geneSymbol_action()} Returns the gene symbol from a single Gene id. accession.
}
\seealso{
\code{\link{refseq_protein2RNA}} to obtain the XM ids that encode a set of XP ids.

\code{\link{refseq_RNA2protein}} to obtain the XP ids encoded by a set of XM ids.
}
\author{
Jose V. Die
}
