% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refseqR.R
\docType{package}
\name{refseqR}
\alias{refseqR}
\alias{refseqR-package}
\title{refseqR: Common computational operations working with RefSeq}
\description{
refseqR is a framework of common computational operations working with RefSeq entries (GenBank)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jdieramon/refseqR}
  \item Report bugs at \url{https://github.com/jdieramon/refseqR/issues}
}

}
\author{
Jose V. Die \email{jose.die@uco.es}
}
