% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_CDSfasta.R
\name{save_CDSfasta_from_xms}
\alias{save_CDSfasta_from_xms}
\title{Save the CDS nucleotide sequence into a file}
\usage{
save_CDSfasta_from_xms(xmsIds, nameFile)
}
\arguments{
\item{xmsIds}{a vector containing a single/multiple XM ids.}

\item{nameFile}{the name of the output file (do not include "fasta").}
}
\description{
Parses a single/multiple XM acessions (Genbank format) and save
the CDS nucleotide sequences into a single / multiple fasta file.
}
\author{
Jose V. Die
}
