% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_AAfasta.R
\name{save_AAfasta_from_xps}
\alias{save_AAfasta_from_xps}
\title{Save the amino acid sequence into a file}
\usage{
save_AAfasta_from_xps(xpsIds, nameFile)
}
\arguments{
\item{xpsIds}{a vector containing a single/multiple XP ids.}

\item{nameFile}{the name of the output file (do not include "fasta").}
}
\description{
Parses a single/multiple XP acessions (Genbank format) and save
the amino acid sequences into a single / multiple fasta file.
}
\examples{
\dontrun{
# xps = c("XP_020271897", "XP_020271898", "XP_020271899")
# save_AAfasta_from_xps(xps, "my_AA")
}

}
\author{
Jose V. Die
}
